refacto: dispatch frontend types into several type files
