+++ packages/web/src/types/envTypes.ts
@@ -1,0 +1,9 @@
+// FOR ENV 
+
+export interface MetaEnv {
+  VITE_DATA_DIR_PATH: string,
+  VITE_DEPLOY_URL: string,
+  VITE_BREVO_TOKEN: string,
+  VITE_BREVO_LIST_IDS: string,
+  VITE_BREVO_SENDER_EMAIL: string,
+}
