+++ packages/web/src/types/translationTypes.ts
@@ -1,0 +1,11 @@
+// FOR TRANSLATIONS
+
+export interface Translations {
+  fr: string,
+  [name: string]: string
+}
+
+export interface Dict {
+  id: string,
+  label: any,
+}
